+++ co-log-polysemy/CHANGELOG.md
@@ -3,6 +3,11 @@
 `co-log-polysemy` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
+## Unreleased
+
+* [#121](https://github.com/kowainik/co-log/issues/121):
+  Add `LogAction` and interpreter that works directly with the `Sem` monad.
+
 ## 0.0.0.0 — May 5, 2019
 
 * Initially created.